Fix YouTube download library

+++ get_youtube.py
@@ -1,13 +1,9 @@
-import youtube_dl
+import yt_dlp
 import os
 import sys
 import json
 
 if __name__ == '__main__':
-
-    yt_id = []
-    data_id = []
-    offset_list = []
 
     yt_link_path = sys.argv[1]
     train_set_dir = sys.argv[2]
@@ -22,7 +18,11 @@ if __name__ == '__main__':
     with open(yt_link_path, newline='') as file:
         yt_link = json.load(file)
 
-    # song #1~#400 are training set 
+    # For easier download tracking
+    success = 0
+    failed = 0
+
+    # Songs #1~#400 are training set 
     for i in range(1, 401):
 
         out_dir = os.path.join(train_set_dir, str(i))
@@ -30,9 +30,11 @@ if __name__ == '__main__':
         if not os.path.isdir(out_dir):
             os.mkdir(out_dir)
 
-        outtmpl = os.path.join(out_dir, "Mixture.mp3")
+        outtmpl = os.path.join(out_dir, "Mixture")
 
         if os.path.isfile(outtmpl):
+            print("Song id", i, "already downloaded, skipping")
+            success += 1
             continue
 
         ydl_opts = {
@@ -44,19 +46,22 @@ if __name__ == '__main__':
                 "preferredquality": "320"
             }],
             "quiet": True,
+            "no_warnings": True,
         }
 
         try:
-            with youtube_dl.YoutubeDL(ydl_opts) as ydl:
-                youtube_dl.utils.std_headers['User-Agent'] = "Mozilla/5.0 (compatible; Googlebot/2.1; +http://www.google.com/bot.html)"
+            with yt_dlp.YoutubeDL(ydl_opts) as ydl:
                 ydl.download([yt_link[str(i)]])
-                print ("Song id", i, "OK")
-        except:
-            print ("Song id", i, "not available")
-            os.rmdir(out_dir)
+                print("Song id", i, "OK")
+                success += 1
+        except Exception as e:
+            print("Song id", i, "not available -", str(e)[:60])
+            failed += 1
+            # Remove the created empty directory if download failed
+            if os.path.isdir(out_dir) and not os.listdir(out_dir):
+                os.rmdir(out_dir)
 
-
-    # song #401~#500 are test set 
+    # Songs #401~#500 are test set 
     for i in range(401, 501):
 
         out_dir = os.path.join(test_set_dir, str(i))
@@ -64,9 +69,11 @@ if __name__ == '__main__':
         if not os.path.isdir(out_dir):
             os.mkdir(out_dir)
 
-        outtmpl = os.path.join(out_dir, "Mixture.mp3")
+        outtmpl = os.path.join(out_dir, "Mixture")
 
         if os.path.isfile(outtmpl):
+            print("Song id", i, "already downloaded, skipping")
+            success += 1
             continue
 
         ydl_opts = {
@@ -78,12 +85,18 @@ if __name__ == '__main__':
                 "preferredquality": "320"
             }],
             "quiet": True,
+            "no_warnings": True,
         }
 
         try:
-            with youtube_dl.YoutubeDL(ydl_opts) as ydl:
-                youtube_dl.utils.std_headers['User-Agent'] = "Mozilla/5.0 (compatible; Googlebot/2.1; +http://www.google.com/bot.html)"
+            with yt_dlp.YoutubeDL(ydl_opts) as ydl:
                 ydl.download([yt_link[str(i)]])
-                print ("Song id", i, "OK")
-        except:
-            print ("Song id", i, "not available")
+                print("Song id", i, "OK")
+                success += 1
+        except Exception as e:
+            print("Song id", i, "not available -", str(e)[:60])
+            failed += 1
+            if os.path.isdir(out_dir) and not os.listdir(out_dir):
+                os.rmdir(out_dir)
+
+    print(f"\nDone! {success} downloaded, {failed} failed")
